requirements.txt
streamlit>=1.12.0
openai>=0.27.0
python-dotenv>=0.20.0